version 26.05.26.20.33 - phase4 extract SiteConfigManager and add gate evidence
