+++ specs/193-main-decomposition-wave-2/checklists/phase-4-gate.md
@@ -1,0 +1,48 @@
+# Phase 4 Gate Evidence
+
+Date: 2026-05-26
+Scope: T028/T029/T030/T031/T031A/T031B/T032/T033/T034/T035/T036
+
+## Code Extraction and Delegation
+
+- Extracted `SiteConfigManager` implementation from `MistHelper.py` into `src/site/site_config_manager.py`.
+- Updated `MistHelper.py` `SiteConfigManager` to orchestration/delegation wrapper for menu operations `171-174`.
+- Added `src/site/__init__.py` and unit test coverage at `tests/unit/site/test_site_config_manager.py`.
+- No changes were made to `GlobalImportManager`.
+
+## Mandatory Validation Commands (T031)
+
+### `python -m py_compile MistHelper.py`
+
+- Result: pass (no output, exit code 0).
+
+### `python -m ruff check MistHelper.py`
+
+- Result: `All checks passed!`.
+
+### `python -m black --check MistHelper.py`
+
+- Result: `1 file would be left unchanged.`
+
+### `python -m pytest tests/unit/site/test_site_config_manager.py tests/contract/test_import_graph.py tests/integration/test_runtime_coupling.py -q`
+
+- Result: `19 passed, 1 warning in 2.02s`.
+
+## Constitution Compliance (T031A)
+
+- `MistHelper.py` now contains delegation-only entrypoints for SiteConfigManager menu operations `171-174`.
+- Runtime dependency wiring is explicit from `MistHelper.py` into `src/site/site_config_manager.py`.
+- Scope guard respected: `GlobalImportManager` unchanged.
+
+## Full Deployment Pipeline Attempt (T031B)
+
+- Status: **Pending update in this run section after commit/push + CI wait + image pull + runtime verification**.
+
+## Import Graph and Runtime Coupling Gates
+
+- `tests/contract/test_import_graph.py`: pass (T034).
+- `tests/integration/test_runtime_coupling.py`: pass, includes `phase_4` profile selection checks (T035).
+
+## Phase 4 Signoff (T036)
+
+- Pending final signoff update after deployment pipeline attempt (T031B) is recorded.
