+++ specs/193-main-decomposition-wave-2/checklists/phase-4-menu-parity.md
@@ -1,0 +1,32 @@
+# Phase 4 Menu Parity Evidence (Operations 171-174)
+
+Date: 2026-05-26
+
+## Operation 171 - Create test sites from CSV
+
+- Menu entry remains option `171` with unchanged description and destructive classification.
+- `MistHelper.py` now delegates `SiteConfigManager.create_test_sites_from_csv()` into `src/site/site_config_manager.py`.
+- Confirmation keyword (`CREATE`) and output artifact flow remain unchanged.
+
+## Operation 172 - Create country RF templates and assign
+
+- Menu entry remains option `172` with unchanged description and destructive classification.
+- `MistHelper.py` delegates to extracted implementation for RF template analysis/create/update/assign flow.
+- Existing update-mode prompt (`1` skip / `2` update) and typed confirmation (`CREATE`) are preserved.
+
+## Operation 173 - Create AP model device profiles
+
+- Menu entry remains option `173` with unchanged description and destructive classification.
+- `MistHelper.py` delegates to extracted implementation for AP model scan and profile creation.
+- User confirmation phrase (`CREATE`) and report output behavior are preserved.
+
+## Operation 174 - Assign APs to matching device profiles
+
+- Menu entry remains option `174` with unchanged description and destructive classification.
+- `MistHelper.py` delegates to extracted implementation for AP/profile matching and assignment.
+- Confirmation keyword (`ASSIGN`) and CSV reporting behavior are preserved.
+
+## Conclusion
+
+- Menu IDs, dispatch keys, and user-facing descriptions for `171-174` are unchanged.
+- Runtime behavior is preserved while canonical implementation ownership moved from `MistHelper.py` to `src/site/site_config_manager.py`.
