+++ specs/193-main-decomposition-wave-2/checklists/phase-4-output-parity.md
@@ -1,0 +1,35 @@
+# Phase 4 Output Parity Evidence (API/Backend)
+
+Date: 2026-05-26
+
+## Scope
+
+- Operation `171` output artifacts:
+  - `CreatedTestSites.csv`
+  - `FailedTestSites.csv`
+- Operation `172` output artifacts:
+  - `SuccessfulRFTemplateAssignments.csv`
+  - `FailedRFTemplateAssignments.csv`
+- Operation `173` output artifacts:
+  - `CreatedAPModelDeviceProfiles.csv`
+  - `FailedAPModelDeviceProfiles.csv`
+- Operation `174` output artifacts:
+  - `SuccessfulAPProfileAssignments.csv`
+  - `FailedAPProfileAssignments.csv`
+  - `SkippedAPsNoMatchingProfile.csv`
+
+## Parity Verification Approach
+
+- Canonical implementation moved to `src/site/site_config_manager.py`.
+- Preserved output boundaries through `DataExporter.save_data_to_output(...)`.
+- Preserved API call intent and menu wiring by keeping `MistHelper.py` as orchestration/delegation only.
+
+## Test Gate Evidence
+
+- Executed required gate suite:
+  - `python -m pytest tests/unit/site/test_site_config_manager.py tests/contract/test_import_graph.py tests/integration/test_runtime_coupling.py -q`
+- Result: `19 passed, 1 warning in 2.02s`.
+
+## Conclusion
+
+- Phase 4 extraction preserves output shaping and backend write paths for operations `171-174` within automated gate scope.
