+++ src/site/__init__.py
@@ -1,0 +1,1 @@
+"""Site configuration modules for MistHelper."""
